+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/_client.py
@@ -34,7 +34,7 @@ class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # 
     Summarization and Conversational PII detection.
 
     :param endpoint: Supported Cognitive Services endpoint (e.g.,
-     https://:code:`<resource-name>`.api.cognitiveservices.azure.com). Required.
+     https://:code:`<resource-name>`.cognitiveservices.azure.com). Required.
     :type endpoint: str
     :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials.AzureKeyCredential
