+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/authoring/__init__.py
@@ -6,13 +6,18 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._operations import ConversationAnalysisProjectsClientOperationsMixin
+from ._client import ConversationAuthoringClient
+from ._version import VERSION
 
-from ._patch import __all__ as _patch_all
-from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+__version__ = VERSION
+
+try:
+    from ._patch import __all__ as _patch_all
+    from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+except ImportError:
+    _patch_all = []
 from ._patch import patch_sdk as _patch_sdk
-__all__ = [
-    'ConversationAnalysisProjectsClientOperationsMixin',
-]
+__all__ = ['ConversationAuthoringClient']
 __all__.extend([p for p in _patch_all if p not in __all__])
+
 _patch_sdk()
