+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/authoring/_operations/__init__.py
@@ -6,13 +6,13 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._operations import ConversationAnalysisProjectsClientOperationsMixin
+from ._operations import ConversationAuthoringClientOperationsMixin
 
 from ._patch import __all__ as _patch_all
 from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
 from ._patch import patch_sdk as _patch_sdk
 __all__ = [
-    'ConversationAnalysisProjectsClientOperationsMixin',
+    'ConversationAuthoringClientOperationsMixin',
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
 _patch_sdk()
