+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/authoring/aio/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._client import ConversationAnalysisProjectsClient
+from ._client import ConversationAuthoringClient
 
 try:
     from ._patch import __all__ as _patch_all
@@ -14,7 +14,7 @@ try:
 except ImportError:
     _patch_all = []
 from ._patch import patch_sdk as _patch_sdk
-__all__ = ['ConversationAnalysisProjectsClient']
+__all__ = ['ConversationAuthoringClient']
 __all__.extend([p for p in _patch_all if p not in __all__])
 
 _patch_sdk()
