+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/authoring/aio/_client.py
@@ -14,14 +14,14 @@ from azure.core.credentials import AzureKeyCredential
 from azure.core.rest import AsyncHttpResponse, HttpRequest
 
 from .._serialization import Deserializer, Serializer
-from ._configuration import ConversationAnalysisProjectsClientConfiguration
-from ._operations import ConversationAnalysisProjectsClientOperationsMixin
+from ._configuration import ConversationAuthoringClientConfiguration
+from ._operations import ConversationAuthoringClientOperationsMixin
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from typing import Dict
 
-class ConversationAnalysisProjectsClient(ConversationAnalysisProjectsClientOperationsMixin):  # pylint: disable=client-accepts-api-version-keyword
+class ConversationAuthoringClient(ConversationAuthoringClientOperationsMixin):  # pylint: disable=client-accepts-api-version-keyword
     """The language service API is a suite of natural language processing (NLP) skills built with
     best-in-class Microsoft machine learning algorithms. The API can be used to analyze
     unstructured text for tasks such as sentiment analysis, key phrase extraction, language
@@ -29,7 +29,7 @@ class ConversationAnalysisProjectsClient(ConversationAnalysisProjectsClientOpera
     href="https://docs.microsoft.com/en-us/azure/cognitive-services/language-service/overview">https://docs.microsoft.com/en-us/azure/cognitive-services/language-service/overview</a>`.
 
     :param endpoint: Supported Cognitive Services endpoint (e.g.,
-     https://:code:`<resource-name>`.api.cognitiveservices.azure.com). Required.
+     https://:code:`<resource-name>`.cognitiveservices.azure.com). Required.
     :type endpoint: str
     :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials.AzureKeyCredential
@@ -47,7 +47,7 @@ class ConversationAnalysisProjectsClient(ConversationAnalysisProjectsClientOpera
         **kwargs: Any
     ) -> None:
         _endpoint = '{Endpoint}/language'
-        self._config = ConversationAnalysisProjectsClientConfiguration(endpoint=endpoint, credential=credential, **kwargs)
+        self._config = ConversationAuthoringClientConfiguration(endpoint=endpoint, credential=credential, **kwargs)
         self._client = AsyncPipelineClient(base_url=_endpoint, config=self._config, **kwargs)
 
         self._serialize = Serializer()
@@ -88,7 +88,7 @@ class ConversationAnalysisProjectsClient(ConversationAnalysisProjectsClientOpera
     async def close(self) -> None:
         await self._client.close()
 
-    async def __aenter__(self) -> "ConversationAnalysisProjectsClient":
+    async def __aenter__(self) -> "ConversationAuthoringClient":
         await self._client.__aenter__()
         return self
